requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Markdown==3.3.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.2